Conflicts:
	Applications/AutoCalibration/Main.cpp
	Applications/Calibration/CalibrationClient.cpp
	Applications/Calibration/Main.cpp
	Applications/Calibration/SystemWidget.cpp
	Applications/LaunchStack/Controllable.cpp
	Applications/LaunchStack/DeftModule.cpp
	Applications/LaunchStack/PreviewImageModule.cpp
	Applications/LaunchStack/PreviewImageModule.hpp
	Applications/LaunchStack/RemoteApplication.cpp
	Applications/LaunchStack/RemoteApplication.hpp
	Applications/LaunchStack/TactionModule.cpp
	CornerStone.pri
	CornerStone.pro
	Experimental/DeftViewer/DeftViewer.pro
	MultiStateDisplay/DeftVisualizer.cpp
	MultiTouch/Version.hpp
	MultiWidgets/Application.cpp
	MultiWidgets/Application.hpp
	MultiWidgets/MultiWidgets.pro
	MultiWidgets/SimpleSDLApplication.hpp
	ThreadedRendering/Application.cpp
	ThreadedRendering/SimpleThreadedApplication.cpp
	ThreadedRendering/SimpleThreadedApplication.hpp
	VERSION
	multitude/Applications/FireView/FireView.pro
	multitude/Applications/MoviePlayer/MoviePlayer.pro
	multitude/Luminous/GLResources.hpp
	multitude/Luminous/Luminous.pro
	multitude/Luminous/MultiHead.cpp
	multitude/Luminous/MultiHead.hpp
	multitude/Radiant/QtWindow.cpp
	multitude/Radiant/Radiant.pro
	multitude/Valuable/Node.cpp
	multitude/Valuable/Node.hpp
	multitude/library.pri
	multitude/multitude.pri

+++ multitude/Poetic/Utils.cpp
@@ -156,21 +156,21 @@ namespace Poetic
     if(ws.empty())
       return;
 
-    size_t n = ws.size();
+    int n = (int)ws.size();
 
     std::vector<float> advances;
     advances.resize(n, 0);
 
-    fnt.advanceList(ws.c_str(), & advances[0], static_cast<int> (n));
+    fnt.advanceList(ws.c_str(), & advances[0], n);
 
-    size_t lineStart = 0;
+    int lineStart = 0;
     int okEnd = 0;
     float sum = 0.0f;
     float okEndSum = 0.0f;
 
     bool onspace = false;
 
-    for(size_t i = 0; i < n; i++) {
+    for(int i = 0; i < n; i++) {
 
       int c = ws[i];
       float a = advances[i];
